hash: fix PyCharm autocompletion of passlib.hash
Issue for a while has been that IDEs such as PyCharm can't detect
autocompletion in 'passlib.hash', because the module is replaced
w/ a lazy-loading object, that breaks it's source inspection.

As workaround, that module now contains a block listing all the imports
explicitly.  This block is trapped in a "if False:" block so it will do nothing,
but that seems sufficient for PyCharm to autocomplete with.

Script "admin/regen.py" can be used to autogenerate this list when registry.py changes.