Back out changeset 53dc260cead7 -- restores argon2 to 16 byte hash output.
per followup to original issue (https://bitbucket.org/ecollins/passlib/issues/88#comment-35296567)
it was pointed out that citation for 32 byte hash output was out of date,
and that argon2_cffi is correctly following the more up-to-date argon2 spec.

this restores the stable branch to the 1.7.1 policy of following argon2_cffi's defaults.